requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
python-dotenv==1.0.0
torch==2.0.1 --index-url https://download.pytorch.org/whl/cpu
numpy==1.24.3 